backend/requirements.txt
fastapi
uvicorn
google-genai
chromadb
python-dotenv
pydantic
python-multipart
sentence-transformers
